plugin-client: handle needsIdentity in dx account login --method email
`loginWithEmail` read only `LoginResponse.token`, so for an address the hub is
willing to bind a fresh Account to it printed "paste the token from the email"
and waited for a token that is never sent — the login could not complete.

Adds the branch the gate has in WelcomeScreen.handleLogin: on `needsIdentity`,
create a local identity via ClientOperation.CreateIdentity (which fires
IdentityCreated, provisioning the personal space), retry the login with its DID,
and on admission provision the agent. Fails loudly if the retry is not admitted
rather than falling through to the token prompt, which would leave a dangling
local identity behind.

The token and out-of-band prompt paths are unchanged.
