--- .changeset/dx-account-signup.md
@@ -1,5 +1,0 @@
----
-'@dxos/plugin-client': minor
----
-
-Add `dx account signup --email` for creating a DXOS identity and Hub account from the terminal.
