--- packages/plugins/plugin-client/src/commands/account/signup/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-export * from './signup';
